Move addLambdaDeserialize into the BTypes assembly
This allows accessing the method from the inliner, which doesn't have
access to the Global instance.
