+++ src/compiler/scala/tools/nsc/backend/jvm/BCodeHelpers.scala
@@ -683,60 +683,6 @@ abstract class BCodeHelpers extends BCodeIdiomatic with BytecodeWriters {
         new java.lang.Long(id)
       ).visitEnd()
     }
-
-    /**
-     * Add:
-     * private static java.util.Map $deserializeLambdaCache$ = null
-     * private static Object $deserializeLambda$(SerializedLambda l) {
-     *   var cache = $deserializeLambdaCache$
-     *   if (cache eq null) {
-     *     cache = new java.util.HashMap()
-     *     $deserializeLambdaCache$ = cache
-     *   }
-     *   return scala.runtime.LambdaDeserializer.deserializeLambda(MethodHandles.lookup(), cache, l);
-     * }
-     */
-    def addLambdaDeserialize(clazz: Symbol, jclass: asm.ClassVisitor): Unit = {
-      val cw = jclass
-      import scala.tools.asm.Opcodes._
-
-      // Need to force creation of BTypes for these as `getCommonSuperClass` is called on
-      // automatically computing the max stack size (`visitMaxs`) during method writing.
-      javaUtilHashMapReference
-      javaUtilMapReference
-
-      cw.visitInnerClass("java/lang/invoke/MethodHandles$Lookup", "java/lang/invoke/MethodHandles", "Lookup", ACC_PUBLIC + ACC_FINAL + ACC_STATIC)
-
-      {
-        val fv = cw.visitField(ACC_PRIVATE + ACC_STATIC + ACC_SYNTHETIC, "$deserializeLambdaCache$", "Ljava/util/Map;", null, null)
-        fv.visitEnd()
-      }
-
-      {
-        val mv = cw.visitMethod(ACC_PRIVATE + ACC_STATIC + ACC_SYNTHETIC, "$deserializeLambda$", "(Ljava/lang/invoke/SerializedLambda;)Ljava/lang/Object;", null, null)
-        mv.visitCode()
-        // javaBinaryName returns the internal name of a class. Also used in BTypesFromsymbols.classBTypeFromSymbol.
-        mv.visitFieldInsn(GETSTATIC, clazz.javaBinaryName.toString, "$deserializeLambdaCache$", "Ljava/util/Map;")
-        mv.visitVarInsn(ASTORE, 1)
-        mv.visitVarInsn(ALOAD, 1)
-        val l0 = new asm.Label()
-        mv.visitJumpInsn(IFNONNULL, l0)
-        mv.visitTypeInsn(NEW, "java/util/HashMap")
-        mv.visitInsn(DUP)
-        mv.visitMethodInsn(INVOKESPECIAL, "java/util/HashMap", "<init>", "()V", false)
-        mv.visitVarInsn(ASTORE, 1)
-        mv.visitVarInsn(ALOAD, 1)
-        mv.visitFieldInsn(PUTSTATIC, clazz.javaBinaryName.toString, "$deserializeLambdaCache$", "Ljava/util/Map;")
-        mv.visitLabel(l0)
-        mv.visitFieldInsn(GETSTATIC, "scala/runtime/LambdaDeserializer$", "MODULE$", "Lscala/runtime/LambdaDeserializer$;")
-        mv.visitMethodInsn(INVOKESTATIC, "java/lang/invoke/MethodHandles", "lookup", "()Ljava/lang/invoke/MethodHandles$Lookup;", false)
-        mv.visitVarInsn(ALOAD, 1)
-        mv.visitVarInsn(ALOAD, 0)
-        mv.visitMethodInsn(INVOKEVIRTUAL, "scala/runtime/LambdaDeserializer$", "deserializeLambda", "(Ljava/lang/invoke/MethodHandles$Lookup;Ljava/util/Map;Ljava/lang/invoke/SerializedLambda;)Ljava/lang/Object;", false)
-        mv.visitInsn(ARETURN)
-        mv.visitEnd()
-      }
-    }
   } // end of trait BCClassGen
 
   /* functionality for building plain and mirror classes */
